Refactor into its own function
