+++ bodo/pandas/base.py
@@ -151,8 +151,8 @@ def read_iceberg(
     import pyiceberg.catalog
     import pyiceberg.expressions
     import pyiceberg.table
-    from pyiceberg.manifest import ManifestContent, ManifestEntryStatus
 
+    from bodo.io.iceberg.read_metadata import get_table_length
     from bodo.pandas.utils import BodoLibNotImplementedException
 
     # Support simple directory only calls like:
@@ -186,51 +186,7 @@ def read_iceberg(
     empty_df = arrow_to_empty_df(arrow_schema)
 
     # Get the table length estimate, if there's not a filter it will be exact
-    table_len_estimate = None
-    snapshot = (
-        table.current_snapshot()
-        if snapshot_id is None
-        else table.snapshot_by_id(snapshot_id)
-    )
-    assert snapshot is not None
-    # If the snapshot has a summary with total-records, use that.
-    if (
-        hasattr(snapshot, "summary")
-        and snapshot.summary is not None
-        and "total-records" in snapshot.summary
-    ):
-        table_len_estimate = int(snapshot.summary["total-records"])
-    # If the snapshot has manifests with existing_rows_count and added_rows_count,
-    # use those to get the table length.
-    elif all(
-        hasattr(manifest, "existing_rows_count")
-        and manifest.existing_rows_count is not None
-        and hasattr(manifest, "added_rows_count")
-        and manifest.added_rows_count is not None
-        for manifest in snapshot.manifests(table.io)
-    ):
-        table_len_estimate = sum(
-            [
-                manifest.existing_rows_count + manifest.added_rows_count
-                if manifest.content == ManifestContent.DATA
-                else 0
-                for manifest in snapshot.manifests(table.io)
-            ]
-        )
-    # Otherwise we need to go through the manifest entries to estimate the table length.
-    else:
-        table_len_estimate = 0
-        for manifest in snapshot.manifests(table.io):
-            manifest_entries = manifest.fetch_manifest_entry(table.io)
-            for entry in manifest_entries:
-                if (
-                    entry.status == ManifestEntryStatus.DELETED
-                    or entry.content != ManifestContent.DATA
-                ):
-                    continue
-                datafile = entry.data_file
-                if datafile is not None and datafile.record_count is not None:
-                    table_len_estimate += datafile.record_count
+    table_len_estimate = get_table_length(table, snapshot_id or -1)
 
     # If there's a row filter, we need to estimate the selectivity
     # and adjust the table length estimate accordingly.
